Make AddReason standalone class
